requirements.txt
streamlit>1.32.0
pytesseract>0.3.10
opencv-python>4.8.0
numpy>1.26.0
Pillow>10.0.0
